Fix sys_rfork hang and resurrection server startup - kernel/9front-pc64/irq.c: Move LAPIC EOI before handler loop to prevent deadlock on context switch. - kernel/9front-pc64/trap.c: Explicitly set IF (0x200) in r11/flags on syscall return to ensure userspace interrupts enabled. - kernel/9front-port/fault.c: Move spllo() to end of fault handler to allow proper preemption. - kernel/9p_router.c: Fix capability hash comparison (6 bytes). - Cleanup debug prints.

+++ kernel/wasm/wasm_runtime/wasm3/m3_compile.c
@@ -2927,7 +2927,7 @@ M3Result CompileFunction(IM3Function io_function) {
     u16 maxSlot = GetMaxUsedSlotPlusOne(o);
 
     o->function->numLocalBytes =
-        (maxSlot - o->slotFirstLocalIndex) * sizeof(m3slot_t);
+        (u16)((maxSlot - o->slotFirstLocalIndex) * sizeof(m3slot_t));
 
     o->slotFirstConstIndex = o->slotMaxConstIndex = maxSlot;
